refactor(gamemap): extract logic
