+++ src/main/scala/scatan/model/GameMap.scala
@@ -1,57 +1,22 @@
 package scatan.model
 
-import scatan.utils.UnorderedTriple
-import scatan.utils.UnorderedPair
-import scatan.model.map.HexTiledMap.*
+import scatan.model.map.HexagonInMap.*
 import scatan.model.map.*
 
-/** A Spot is unique identified by three hexagons in the map.
+/** Hexagonal tiled game map of Scatan.
   *
-  * @param hexagons
-  */
-type Spot = UnorderedTriple[Hexagon]
-
-/** A Road connect two spots.
-  */
-type Road = UnorderedPair[Spot]
-
-/** Hexagonal tiled GameMap of Scatan.
+  * @param withTerrainLayers
+  *   number of concentric circles of hexagons of terrains
   *
-  * @param layers
-  *   number of concentric circles of hexagons
+  * @param withSeaLayers
+  *   number of concentric circles of hexagons the terrain ones.
   */
-final case class GameMap(layers: Int)
-    extends HexTiledMap
-    with UndirectedGraph[Spot, Road]
-    with MapWithTerrain
-    with MapWithNumbers:
+final case class GameMap(withTerrainLayers: Int = 2, withSeaLayers: Int = 1)
+    extends HexagonalTiledMap(withTerrainLayers + withSeaLayers)
+    with MapWithTileContent:
 
-  override val toTerrain: Map[Hexagon, Terrain] =
-    TerrainFactory.fixedForLayer2(tiles.toSeq)
+  val totalLayers = withTerrainLayers + withSeaLayers
+  val tileWithTerrain = tiles.toSeq.filter(_.layer <= withTerrainLayers)
 
-  override val possibleNumbers: Seq[Int] =
-    NumbersFactory.fixedForLayer2(tiles.toSeq)
-
-  override def tiles: Set[Hexagon] =
-    (for
-      r <- -layers to layers
-      q <- -layers to layers
-      s <- -layers to layers
-      if r + q + s == 0
-    yield Hexagon(r, q, s)).toSet
-
-  override def nodes: Set[Spot] =
-    for
-      hex <- tiles
-      first <- hex.neighbours
-      second <- hex.neighbours
-      if first.isNeighbour(second)
-    yield UnorderedTriple(hex, first, second)
-
-  val spotsPerRoad = 2
-  override def edges: Set[Road] =
-    for
-      first <- nodes
-      second <- nodes
-      if first.toSet.intersect(second.toSet).sizeIs == spotsPerRoad
-    yield UnorderedPair(first, second)
+  override def toContent: Map[Hexagon, TileContent] =
+    TileContentFactory.fixedForLayer2(tileWithTerrain)
